+++ src/controller/tickets.rs
@@ -9,7 +9,7 @@ use rocket::serde::json::Json;
 #[post("/tickets/<ticket_id>/assignments", data = "<create_assignment_json>")]
 pub async fn create_assignment_by_ticket_id_endpoint(
     conn: TiraDbConn,
-    ticket_id: i32,
+    ticket_id: i64,
     create_assignment_json: Json<CreateAssignmentWithUserId>,
 ) {
     tickets::create_assignment_by_ticket_id(conn, ticket_id, create_assignment_json.0.user_id)
@@ -19,7 +19,7 @@ pub async fn create_assignment_by_ticket_id_endpoint(
 #[post("/tickets/<ticket_id>/comments", data = "<create_comment_json>")]
 pub async fn create_comment_endpoint(
     conn: TiraDbConn,
-    ticket_id: i32,
+    ticket_id: i64,
     create_comment_json: Json<CreateComment>,
 ) {
     tickets::create_comment(
@@ -46,7 +46,7 @@ pub async fn create_ticket_endpoint(
 }
 
 #[delete("/tickets/<ticket_id>")]
-pub async fn delete_ticket_by_id_endpoint(conn: TiraDbConn, ticket_id: i32) {
+pub async fn delete_ticket_by_id_endpoint(conn: TiraDbConn, ticket_id: i64) {
     tickets::delete_ticket_by_id(conn, ticket_id).await;
 }
 
@@ -58,7 +58,7 @@ pub async fn delete_tickets_endpoint(conn: TiraDbConn) {
 #[get("/tickets/<ticket_id>/assignments")]
 pub async fn get_assignments_by_ticket_id_endpoint(
     conn: TiraDbConn,
-    ticket_id: i32,
+    ticket_id: i64,
 ) -> Json<Vec<Assignment>> {
     Json(tickets::get_assignments_by_ticket_id(conn, ticket_id).await)
 }
@@ -66,13 +66,13 @@ pub async fn get_assignments_by_ticket_id_endpoint(
 #[get("/tickets/<ticket_id>/comments")]
 pub async fn get_comments_by_ticket_id_endpoint(
     conn: TiraDbConn,
-    ticket_id: i32,
+    ticket_id: i64,
 ) -> Json<Vec<Comment>> {
     Json(tickets::get_comments_by_ticket_id(conn, ticket_id).await)
 }
 
 #[get("/tickets/<ticket_id>")]
-pub async fn get_ticket_by_id_endpoint(conn: TiraDbConn, ticket_id: i32) -> Json<Ticket> {
+pub async fn get_ticket_by_id_endpoint(conn: TiraDbConn, ticket_id: i64) -> Json<Ticket> {
     Json(tickets::get_ticket_by_id(conn, ticket_id).await)
 }
